# BackendApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BackendApi",
    "description": "Das ist unser erster Dienst",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost:7242",
      "description": "Localhost (Dev)"
    }
  ],
  "paths": {
    "/api/Documents/{id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "operationId": "GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/Documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "operationId": "GetAll",
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentListDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Documents"
        ],
        "operationId": "Add",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentDto"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Documents"
        ],
        "operationId": "Update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentDto"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/Documents/name": {
      "get": {
        "tags": [
          "Documents"
        ],
        "operationId": "Search",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentListDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Documents/query": {
      "post": {
        "tags": [
          "Documents"
        ],
        "operationId": "Query",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryObject"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryObject"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QueryObject"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentListDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DocumentDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "desc": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "hascontent": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "DocumentListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QueryObject": {
        "type": "object",
        "properties": {
          "documentType": {
            "type": "string",
            "nullable": true
          },
          "query": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}